# docs/swagger.yaml
swagger: '2.0'
info:
  title: Living Progress - Build - Mentor Me
  description: Living Progress - Build - Mentor Me API.
  version: "1.0.0"
# the sample made up domain of the service
host: api.mentorme.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
securityDefinitions:
  basicAuth:
    type: basic
# HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
produces:
  - application/json
paths:
...
  /mentees/{id}:
    get:
      tags:
        - Mentee
      summary: get Mentee
      description: This endpoint is responsible for getting Mentee
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the Mentee
          schema:
            $ref: '#/definitions/Mentee'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Mentee
      summary: update Mentee
      description: This endpoint is responsible for updating Mentee
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
        - name: asset
          in: body
          description: The Mentee to update
          schema:
            $ref: '#/definitions/Mentee'
      responses:
        200:
          description: The response containing the Mentee
          schema:
            $ref: '#/definitions/Mentee'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
        - Mentee
      summary: delte Mentee
      description: This endpoint is responsible for deleting Mentee
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /mentees/{id}/averageScore:
    get:
      tags:
        - Mentee
      summary: get Mentee average score
      description: This endpoint is responsible for getting Mentee
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: number
          required: true
      responses:
        200:
          description: The response containing the score
          schema:
            type: number
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /mentees/{id}/matchingMentors:
      get:
        tags:
          - Mentee
        summary: get Mentee match mentors
        description: This endpoint is responsible for getting match mentors Mentee
        security:
          - basicAuth: []
        parameters:
          - name: id
            in: path
            description: The id
            type: number
            required: true
        responses:
          200:
            description: The response containing the match mentors
            schema:
              type: array
              items:
                 $ref: '#/definitions/Mentor'
          400:
            description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
            schema:
              $ref: '#/definitions/Error'
          401:
            description: Authentication Error - ien is missing or invalid
            schema:
              $ref: '#/definitions/Error'
          403:
            description: Access Denied Error - if forbidden to access
            schema:
              $ref: '#/definitions/Error'
          404:
            description: Not Found - if match entity not found
            schema:
              $ref: '#/definitions/Error'
          500:
            description: Internal server error - if the request was properly formatted, but the operation failed on the server side
            schema:
              $ref: '#/definitions/Error'
          default:
            description: Unexpected error.
            schema:
              $ref: '#/definitions/Error'
  /mentees/{id}/confirmParentConsent:
      get:
        tags:
          - Mentee
        summary: confirm parent consent
        description: This endpoint is responsible for confirming parent consent
        security:
          - basicAuth: []
        parameters:
          - name: id
            in: path
            description: The id
            type: number
            required: true
          - name: token
            in: query
            description: The token
            type: string
            required: true
        responses:
          200:
            description: The response containing the result
            schema:
              type: boolean
          400:
            description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
            schema:
              $ref: '#/definitions/Error'
          401:
            description: Authentication Error - ien is missing or invalid
            schema:
              $ref: '#/definitions/Error'
          403:
            description: Access Denied Error - if forbidden to access
            schema:
              $ref: '#/definitions/Error'
          404:
            description: Not Found - if match entity not found
            schema:
              $ref: '#/definitions/Error'
          500:
            description: Internal server error - if the request was properly formatted, but the operation failed on the server side
            schema:
              $ref: '#/definitions/Error'
          default:
            description: Unexpected error.
            schema:
              $ref: '#/definitions/Error'
  /mentees:
    get:
      tags:
        - Mentee
      summary: search for entities
      description: This endpoint is responsible for searching by criteria
      security:
      - basicAuth: []
      parameters:
        - name: institutionId
          in: query
          description: The  institution id
          type: number
          required: false
        - name: status
          in: query
          description: The  status
          type: string
          required: false
        - name: minAveragePerformanceScore
          in: query
          description: The  min average performance scores
          type: string
          required: false
        - name: maxAveragePerformanceScore
          in: query
          description: The  max average performance scores
          type: string
          required: false
        - name: name
          in: query
          description: The name
          type: string
          required: false
        - name: personalInterests
          in: query
          type: number
          required: false
          description: should use personalInterests[0]=id
        - name: professionalInterests
          in: query
          type: number
          required: false
          description: should use professional[0]=id
        - name: assignedToInstitution
          in: query
          description: The assigned to institution flag
          type: boolean
          required: false
        - name: schoolName
          in: query
          description: The school name
          type: string
          required: false
        - name: pageNumber
          in: query
          description: The page number
          type: number
          format: int32
          required: false
        - name: pageSize
          in: query
          description: The page size
          type: number
          format: int32
          required: false
        - name: sortColumn
          in: query
          description: The sort column
          type: string
          required: false
        - name: sortOrder
          in: query
          description: The sort order
          type: string
          required: false
      responses:
        200:
          description: The response containing the InstitutionalPrograms
          schema:
            $ref: '#/definitions/MenteeSearchResult'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
    post:
      tags:
        - Mentee
      summary: create Mentee
      description: This endpoint is responsible for creating Mentee
      security:
        - basicAuth: []
      parameters:
        - name: institution
          in: body
          description: The Mentee to create
          schema:
            $ref: '#/definitions/MenteeCreateBody'
      responses:
        200:
          description: The response containing the Institution
          schema:
            $ref: '#/definitions/Mentee'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  UserRole:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  ParentConsent:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        parentName:
          type: integer
          format: int32
          description: The parent name.
        signatureFilePath:
          type: string
          description: The signature file path.
        parentEmail:
          type: string
          description: The parent email.
        token:
          type: string
          description: The token.
  InstitutionAffiliationCode:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          code:
            type: string
            description: The code.
          institutionId:
            type: number
            description: The institution id.
          used:
            type: boolean
            description: The used flag.
  PersonalInterest:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        value:
          type: string
          description: The value.
        picturePath:
          type: string
          description: The picture path.
        parentCategory:
          $ref: '#/definitions/PersonalInterest'
  WeightedPersonalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          personalInterest:
            $ref: '#/definitions/PersonalInterest'
...
  WeightedProfessionalInterest:
        type: object
        properties:
          id:
            type: integer
            format: int32
            description: The id.
          weight:
            type: number
            description: The weight.
          userId:
            type: number
            description: The user id.
          professionalInterest:
            $ref: '#/definitions/ProfessionalInterest'
...
  Institution:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      institutionName:
        type: string
        description: The name.
      parentOrganization:
        type: string
        description: The parent organization.
      streetAddress:
        type: string
        description: The street address.
      city:
        type: string
        description: The city.
      zip:
        type: string
        description: The zip.
      phone:
         type: string
         description: The phone.
      email:
        type: string
        description: The email.
      description:
        type: string
        description: The description.
      status:
        type: string
        description: The status.
      logoPath:
        type: string
        description: The logo path.
      contacts:
        type: array
        items:
          $ref: '#/definitions/InstitutionContact'
      state:
        $ref: '#/definitions/State'
      country:
        $ref: '#/definitions/Country'
...
  Mentor:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      mentorType:
        type: string
        description: The mentor type.
      companyName:
        type: string
        description: The company name.
      linkedInUrl:
        type: string
        description: The linkedIn url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
...
  MenteeCreateBody:
    type: object
    properties:
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      familyIncome:
        type: number
        description: The family income.
      school:
        type: string
        description: The school.
      facebookUrl:
        type: string
        description: The facebook url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
  Mentee:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      username:
        type: string
        description: The username.
      firstName:
        type: string
        description: The first name.
      lastName:
        type: string
        description: The last name.
      email:
        type: string
        description: The email.
      profilePicturePath:
        type: string
        description: The profile picture path.
      status:
        type: string
        description: The status.
      assignedToInstitution:
        type: boolean
        description: The assigned to institution flag.
      birthDate:
        type: string
        description: The birth date.
      phone:
         type: string
         description: The phone.
      skypeUsername:
        type: string
        description: The skype username.
      introVideoLink:
        type: string
        description: The intro video link.
      description:
        type: string
        description: The description.
      averagePerformanceScore:
        type: number
        description: The average performance score.
      familyIncome:
        type: number
        description: The family income.
      school:
        type: string
        description: The school.
      facebookUrl:
        type: string
        description: The facebook url.
      roles:
        type: array
        items:
          $ref: '#/definitions/UserRole'
      personalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedPersonalInterest'
      professionalInterests:
        type: array
        items:
          $ref: '#/definitions/WeightedProfessionalInterest'
      institution:
        $ref: '#/definitions/Institution'
      institutionAffiliationCode:
        $ref: '#/definitions/InstitutionAffiliationCode'
      parentConsent:
        $ref: '#/definitions/ParentConsent'
  MenteeSearchResult:
    type: object
    properties:
      total:
        type: number
        format: integer
        description: The total.
      totalPages:
        type: number
        format: integer
        description: The total pages.
      entities:
        type: array
        items:
          $ref: '#/definitions/Mentee'
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string